+++ genesyscloud/telephony_providers_edges_edge_group/resource_genesyscloud_telephony_providers_edges_edge_group.go
@@ -1,4 +1,4 @@
-package telephony
+package telephony_providers_edges_edge_group
 
 import (
 	"context"
